# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  trips.Activity:
    properties:
      activity_type:
        allOf:
        - $ref: '#/definitions/trips.ActivityType'
        example: sightseeing
      actual_cost:
        example: 30
        type: number
      booking_ref:
        example: EIF123456
        type: string
      contact_info:
        example: +33 1 44 11 23 23
        type: string
      created_at:
        type: string
      description:
        example: Take elevator to the top, enjoy views
        type: string
      duration:
        example: 120
        type: integer
      end_time:
        example: "2024-06-01T12:00:00Z"
        type: string
      estimated_cost:
        example: 25.5
        type: number
      id:
        type: string
      location:
        example: Champ de Mars, 5 Avenue Anatole France
        type: string
      map_source:
        example: google
        type: string
      name:
        example: Visit Eiffel Tower
        type: string
      notes:
        example: Book tickets in advance to avoid queue
        type: string
      place_id:
        example: ChIJLU7jMh9u5kcR4PcOOO6p3I0
        type: string
      priority:
        example: 1
        type: integer
      start_time:
        example: "2024-06-01T10:00:00Z"
        type: string
      status:
        example: planned
        type: string
      tags:
        example:
        - must-see
        - photo-op
        items:
          type: string
        type: array
      trip_day:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      trip_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
    type: object
  trips.ActivityType:
    enum:
    - transport
    - sightseeing
    - dining
    - shopping
    - entertainment
    - adventure
    - cultural
    - business
    - personal
    - other
    type: string
    x-enum-comments:
      ActivityTypeAdventure: Adventure sports, hiking
      ActivityTypeBusiness: Work-related activities
      ActivityTypeCultural: Museums, historical sites
      ActivityTypeDining: Restaurants, cafes
      ActivityTypeEntertainment: Shows, movies, etc.
      ActivityTypeOther: Other activities
      ActivityTypePersonal: Personal time, rest
      ActivityTypeShopping: Shopping activities
      ActivityTypeSightseeing: Tourist attractions
      ActivityTypeTransport: Transportation between places
    x-enum-descriptions:
    - Transportation between places
    - Tourist attractions
    - Restaurants, cafes
    - Shopping activities
    - Shows, movies, etc.
    - Adventure sports, hiking
    - Museums, historical sites
    - Work-related activities
    - Personal time, rest
    - Other activities
    x-enum-varnames:
    - ActivityTypeTransport
    - ActivityTypeSightseeing
    - ActivityTypeDining
    - ActivityTypeShopping
    - ActivityTypeEntertainment
    - ActivityTypeAdventure
    - ActivityTypeCultural
    - ActivityTypeBusiness
    - ActivityTypePersonal
    - ActivityTypeOther
...
  trips.ContactRole:
    enum:
    - hotel
    - guide
    - driver
    - embassy
    - insurance
    - emergency
    - other
    type: string
    x-enum-varnames:
    - ContactRoleHotel
    - ContactRoleGuide
    - ContactRoleDriver
    - ContactRoleEmbassy
    - ContactRoleInsurance
    - ContactRoleEmergency
    - ContactRoleOther
...
  trips.Currency:
    enum:
    - INR
    - USD
    - GBP
    - EUR
    - CAD
    - AUD
    - JPY
    - OTHER
    type: string
    x-enum-comments:
      CurrencyAUD: Australian Dollar
      CurrencyCAD: Canadian Dollar
      CurrencyEUR: Euro
      CurrencyGBP: British Pound
      CurrencyINR: Indian Rupee
      CurrencyJPY: Japanese Yen
      CurrencyOther: Other currencies
      CurrencyUSD: US Dollar
    x-enum-descriptions:
    - Indian Rupee
    - US Dollar
    - British Pound
    - Euro
    - Canadian Dollar
    - Australian Dollar
    - Japanese Yen
    - Other currencies
    x-enum-varnames:
    - CurrencyINR
    - CurrencyUSD
    - CurrencyGBP
    - CurrencyEUR
    - CurrencyCAD
    - CurrencyAUD
    - CurrencyJPY
    - CurrencyOther
...
  trips.ShiftDatesRequest:
    properties:
      delta_days:
        example: 3
        type: integer
    required:
    - delta_days
    type: object
...
  trips.StatusUpdateRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/trips.TripStatus'
        example: confirmed
    required:
    - status
    type: object
  trips.Stay:
    properties:
      created_at:
        type: string
      end_date:
        example: "2024-06-03T00:00:00Z"
        type: string
      google_location:
        example: ChIJD7fiBh9u5kcRYJSMaMOCCwQ
        type: string
      id:
        type: string
      is_prepaid:
        example: true
        type: boolean
      mapbox_location:
        example: paris.hotel.123
        type: string
      payment_mode:
        example: credit_card
        type: string
      start_date:
        example: "2024-06-01T00:00:00Z"
        type: string
      stay_notes:
        example: Near Eiffel Tower
        type: string
      stay_type:
        example: hotel
        type: string
      trip_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
    type: object
  trips.TodayView:
    properties:
      check_in_today:
        type: boolean
      check_out_today:
        type: boolean
      current_hop:
        $ref: '#/definitions/trips.TripHop'
      current_stay:
        $ref: '#/definitions/trips.Stay'
      date:
        type: string
      day_number:
        type: integer
      emergency_contacts:
        items:
          $ref: '#/definitions/trips.TripContact'
        type: array
      emergency_number:
        type: string
      trip_day:
        $ref: '#/definitions/trips.TripDay'
      trip_id:
        type: string
      trip_name:
        type: string
      upcoming_transport:
        $ref: '#/definitions/trips.TransportSegment'
    type: object
  trips.TransportMode:
    enum:
    - flight
    - train
    - bus
    - car_rental
    - ferry
    - taxi
    - other
    type: string
    x-enum-varnames:
    - TransportModeFlight
    - TransportModeTrain
    - TransportModeBus
    - TransportModeCarRental
    - TransportModeFerry
    - TransportModeTaxi
    - TransportModeOther
  trips.TransportSegment:
    properties:
      arrive_at:
        type: string
      arrive_to:
        example: Mumbai CSMT
        type: string
      arrive_tz:
        example: Asia/Kolkata
        type: string
      booking_ref:
        example: PNR1234567
        type: string
      cost:
        example: 1500
        type: number
      created_at:
        type: string
      currency:
        example: INR
        type: string
      deep_link_url:
        example: https://www.irctc.co.in/nget/train-search
        type: string
      depart_at:
        type: string
      depart_from:
        example: New Delhi Railway Station
        type: string
      depart_tz:
        example: Asia/Kolkata
        type: string
      from_hop_id:
        type: string
      id:
        type: string
      mode:
        allOf:
        - $ref: '#/definitions/trips.TransportMode'
        example: train
      notes:
        type: string
      operator:
        example: IRCTC
        type: string
      seats:
        example: S4 42, S4 43
        type: string
      to_hop_id:
        type: string
      trip_plan:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
    type: object
...
  trips.Traveller:
    properties:
      created_at:
        type: string
      date_of_birth:
        example: "1990-01-15T00:00:00Z"
        type: string
      dietary_restrictions:
        example: Vegetarian, No nuts
        type: string
      email:
        example: john.doe@example.com
        type: string
      emergency_contact:
        example: 'Jane Doe: +1-555-9876'
        type: string
      first_name:
        example: John
        type: string
      id:
        type: string
      is_active:
        type: boolean
      joined_at:
        type: string
      last_name:
        example: Doe
        type: string
      medical_notes:
        example: Allergic to shellfish
        type: string
      nationality:
        example: US
        type: string
      notes:
        example: Prefers window seats
        type: string
      passport_expiry:
        example: "2030-01-15T00:00:00Z"
        type: string
      passport_number:
        example: A12345678
        type: string
      phone:
        example: +1-555-0123
        type: string
      role:
        example: organizer
        type: string
      trip_plan:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  trips.TripContact:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        example: reservations@tajhotels.com
        type: string
      hop_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        example: Taj Hotel Front Desk
        type: string
      notes:
        type: string
      phone:
        example: +91-22-66651234
        type: string
      role:
        allOf:
        - $ref: '#/definitions/trips.ContactRole'
        example: hotel
      trip_plan:
        type: string
      updated_at:
        type: string
    type: object
...
  trips.TripDay:
    properties:
      activities:
        items:
          $ref: '#/definitions/trips.Activity'
        type: array
      actual_budget:
        example: 175.25
        type: number
      created_at:
        type: string
      date:
        example: "2024-06-01"
        format: date
        type: string
      day_number:
        example: 1
        type: integer
      day_type:
        allOf:
        - $ref: '#/definitions/trips.TripDayType'
        example: explore
      end_location:
        example: Eiffel Tower
        type: string
      estimated_budget:
        example: 150.5
        type: number
      from_trip_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      id:
        type: string
      notes:
        example: Start early to avoid crowds
        type: string
      start_location:
        example: Hotel de Paris
        type: string
      title:
        example: Exploring Paris
        type: string
      to_trip_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      trip_plan:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
      weather:
        example: Sunny, 22°C
        type: string
    type: object
  trips.TripDayType:
    enum:
    - travel
    - explore
    - relax
    - business
    - adventure
    - cultural
    type: string
    x-enum-comments:
      TripDayTypeAdventure: Adventure activities
      TripDayTypeBusiness: Business activities
      TripDayTypeCultural: Cultural experiences
      TripDayTypeExplore: Day for exploring/sightseeing
      TripDayTypeRelax: Rest/leisure day
      TripDayTypeTravel: Day primarily for traveling between locations
    x-enum-descriptions:
    - Day primarily for traveling between locations
    - Day for exploring/sightseeing
    - Rest/leisure day
    - Business activities
    - Adventure activities
    - Cultural experiences
    x-enum-varnames:
    - TripDayTypeTravel
    - TripDayTypeExplore
    - TripDayTypeRelax
    - TripDayTypeBusiness
    - TripDayTypeAdventure
    - TripDayTypeCultural
...
  trips.TripHop:
    properties:
      activities:
        example:
        - Seine River Cruise
        - Walking Tour
        items:
          type: string
        type: array
      actual_spent:
        example: 750.5
        type: number
      city:
        example: Paris
        type: string
      country:
        example: France
        type: string
      created_at:
        type: string
      description:
        example: 3 days exploring the City of Light
        type: string
      end_date:
        example: "2024-06-03T00:00:00Z"
        type: string
      estimated_budget:
        example: 800
        type: number
      hop_order:
        example: 1
        type: integer
      id:
        type: string
      latitude:
        example: 48.8566
        type: number
      longitude:
        example: 2.3522
        type: number
      map_source:
        example: google
        type: string
      name:
        example: Paris Visit
        type: string
      next_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      notes:
        example: Visit Eiffel Tower
        type: string
      place_id:
        example: ChIJD7fiBh9u5kcRYJSMaMOCCwQ
        type: string
      pois:
        example:
        - Eiffel Tower
        - Louvre Museum
        items:
          type: string
        type: array
      previous_hop:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      region:
        example: Île-de-France
        type: string
      restaurants:
        example:
        - Le Jules Verne
        - L'As du Fallafel
        items:
          type: string
        type: array
      route_to_next:
        type: object
      selected_flight:
        type: object
      selected_hotel:
        type: object
      start_date:
        example: "2024-06-01T00:00:00Z"
        type: string
      stays:
        items:
          $ref: '#/definitions/trips.Stay'
        type: array
      transportation:
        example: train
        type: string
      trip_days:
        items:
          $ref: '#/definitions/trips.TripDay'
        type: array
      trip_plan:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      updated_at:
        type: string
    type: object
  trips.TripPlan:
    properties:
      actual_spent:
        example: 4750.25
        type: number
      budget:
        example: 5000
        type: number
      created_at:
        type: string
      currency:
        allOf:
        - $ref: '#/definitions/trips.Currency'
        example: EUR
      description:
        example: A romantic 10-day getaway to France
        type: string
      end_date:
        example: "2024-06-10T00:00:00Z"
        type: string
      hotels:
        example:
        - Hotel de Paris
        - Le Bristol
        items:
          type: string
        type: array
      id:
        type: string
      is_public:
        example: false
        type: boolean
      max_days:
        example: 14
        type: integer
      min_days:
        example: 7
        type: integer
      name:
        example: Trip to Paris
        type: string
      notes:
        example: Romantic getaway
        type: string
      participants:
        example:
        - john@example.com
        - jane@example.com
        items:
          type: string
        type: array
      share_code:
        example: PARIS2024ABC
        type: string
      start_date:
        example: "2024-06-01T00:00:00Z"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/trips.TripStatus'
        example: planning
      tags:
        example:
        - romantic
        - europe
        items:
          type: string
        type: array
      timezone:
        example: Asia/Kolkata
        type: string
      travel_modes:
        example:
        - flight
        - train
        items:
          type: string
        type: array
      travellers:
        items:
          $ref: '#/definitions/trips.Traveller'
        type: array
      trip_days:
        items:
          $ref: '#/definitions/trips.TripDay'
        type: array
      trip_hops:
        items:
          $ref: '#/definitions/trips.TripHop'
        type: array
      trip_type:
        example: leisure
        type: string
      updated_at:
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      user_set_status:
        type: boolean
    type: object
  trips.TripStatus:
    enum:
    - planning
    - confirmed
    - ongoing
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - TripStatusPlanning
    - TripStatusConfirmed
    - TripStatusOngoing
    - TripStatusCompleted
    - TripStatusCancelled
host: localhost:8080
info:
  contact: {}
  description: API for managing trip plans, hops, and stays
  title: Trip Planner API
  version: "1.0"
paths:
...
  /trip/{id}/shift-dates:
    post:
      consumes:
      - application/json
      description: Atomically shifts trip start/end, all hop dates, all day dates,
        and all activity times.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: Delta in days (positive or negative)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.ShiftDatesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: days shifted
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Shift all dates in a trip by a number of days
      tags:
      - lifecycle
  /trip/{id}/status:
    patch:
      consumes:
      - application/json
      description: Transitions the trip through its lifecycle. Validates allowed transitions.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/trips.StatusUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated trip plan
          schema:
            $ref: '#/definitions/trips.TripPlan'
        "400":
          description: Invalid transition
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update the status of a trip plan
      tags:
      - lifecycle
  /trip/{id}/today:
    get:
      description: Returns the current day view including current hop/stay, activities,
        upcoming transport, and emergency contacts.
      parameters:
      - description: Trip Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Today's view
          schema:
            $ref: '#/definitions/trips.TodayView'
        "404":
          description: Not found or trip not active
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get today's itinerary for an ongoing trip
      tags:
      - lifecycle
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"